Making plugin.JobStep handle serialization
This way, the JobStep is the only thing that needs to know about the weird way it handles serialization. The JobStepper is the interface and is the only thing on the JobStep. Since it's a component it can be used exactly the same way aside from when it has to be cast to a different type. Not too bad!